Extract shared trajectory message contract
